HZN-1203: Split experimental Karaf features into separate repo
