+++ container/features/pom.xml
@@ -233,14 +233,6 @@
             <feature>jrobin</feature>
             <feature>json-lib</feature>
             <feature>lmax-disruptor</feature>
-            <!--
-            <feature>opennms-activemq-config</feature>
-            <feature>opennms-activemq</feature>
-            <feature>opennms-activemq-dispatcher-config</feature>
-            <feature>opennms-activemq-dispatcher</feature>
-            <feature>opennms-activemq-event-forwarder</feature>
-            <feature>opennms-activemq-event-receiver</feature>
-            -->
             <feature>opennms-bootstrap</feature>
             <feature>opennms-bundle-refresher</feature>
             <feature>opennms-collection-api</feature>
@@ -256,9 +248,7 @@
             <feature>opennms-core-web</feature>
             <feature>opennms-dao-api</feature>
             <feature>opennms-dao</feature>
-            <feature>opennms-discovery</feature>
             <feature>opennms-events-api</feature>
-            <feature>opennms-events-daemon</feature>
             <feature>opennms-events-commands</feature>
             <feature>opennms-icmp-api</feature>
             <feature>opennms-icmp-jna</feature>
@@ -273,7 +263,6 @@
             <feature>opennms-provisioning</feature>
             <feature>opennms-provisioning-detectors</feature>
             <feature>opennms-provisioning-shell</feature>
-            <feature>opennms-reporting</feature>
             <feature>opennms-rrd-api</feature>
             <feature>opennms-rrd-jrobin</feature>
             <feature>opennms-snmp</feature>
@@ -282,7 +271,6 @@
             <feature>opennms-syslogd-listener-javanet</feature>
             <feature>opennms-syslogd-listener-camel-netty</feature>
             <feature>opennms-trapd</feature>
-            <!-- <feature>opennms-webapp</feature> -->
 
             <!-- OSGI Rest -->
             <feature>jax-rs-connector</feature>
@@ -397,6 +385,11 @@
             </goals>
             <configuration>
               <artifacts>
+                <artifact>
+                  <file>${project.build.outputDirectory}/features-experimental.xml</file>
+                  <type>xml</type>
+                  <classifier>experimental</classifier>
+                </artifact>
                 <artifact>
                   <file>${project.build.outputDirectory}/features-minion.xml</file>
                   <type>xml</type>
